CC: Support fractional iteration count on compositor animations
BUG=398767

Review URL: https://codereview.chromium.org/423373003

Cr-Commit-Position: refs/heads/master@{#288931}
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@288931 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 564f12accf9493c73e9cd2926329704ecdd4aeb1

+++ cc/animation/animation.h
@@ -68,8 +68,8 @@ class CC_EXPORT Animation {
   // This is the number of times that the animation will play. If this
   // value is zero the animation will not play. If it is negative, then
   // the animation will loop indefinitely.
-  int iterations() const { return iterations_; }
-  void set_iterations(int n) { iterations_ = n; }
+  double iterations() const { return iterations_; }
+  void set_iterations(double n) { iterations_ = n; }
 
   base::TimeTicks start_time() const { return start_time_; }
 
@@ -161,7 +161,7 @@ class CC_EXPORT Animation {
 
   TargetProperty target_property_;
   RunState run_state_;
-  int iterations_;
+  double iterations_;
   base::TimeTicks start_time_;
   Direction direction_;
